refactor(gsd): split auto-prompts.ts (1621 lines) into 7 focused modules
The monolithic auto-prompts.ts was 1621 lines mixing file inlining utilities,
prompt section builders, context helpers, and milestone/slice/task prompt
constructors. Split into purpose-aligned modules with auto-prompts.ts reduced
to a 16-line barrel re-export.

New modules extracted:
- auto-file-inlining.ts    — inlineFile, inlineFileOptional, inlineFileSmart,
                             inlineDependencySummaries, inlineGsdRootFile,
                             inlineDecisionsFromDb, inlineRequirementsFromDb,
                             inlineProjectFromDb
- auto-prompt-sections.ts  — extractMarkdownSection, escapeRegExp,
                             buildResumeSection, buildCarryForwardSection,
                             extractSliceExecutionExcerpt, capPreamble,
                             buildSourceFilePaths
- auto-prompt-context.ts   — getPriorTaskSummaryPaths, getDependencyTaskSummaryPaths,
                             checkNeedsReassessment, checkNeedsRunUat
- auto-skill-activation.ts — buildSkillActivationBlock, buildSkillDiscoveryVars
- auto-milestone-prompts.ts — buildResearchMilestonePrompt, buildPlanMilestonePrompt,
                              buildCompleteMilestonePrompt, buildValidateMilestonePrompt
- auto-slice-prompts.ts    — buildResearchSlicePrompt, buildPlanSlicePrompt,
                             buildCompleteSlicePrompt, buildReplanSlicePrompt
- auto-task-prompts.ts     — buildExecuteTaskPrompt, buildRunUatPrompt,
                             buildReassessRoadmapPrompt, buildReactiveExecutePrompt,
                             buildRewriteDocsPrompt, ExecuteTaskPromptOptions

Updated triage-dispatch.test.ts to search across split prompt module files
for structural invariants (loadReplanCaptures, loadDeferredCaptures) that moved
from the monolith into auto-slice-prompts.ts and auto-task-prompts.ts.

All imports remain stable via re-exports — no call sites changed.
Typecheck and full unit test suite pass with zero regressions.
